+++ pynsee/utils/init_conn.py
@@ -8,11 +8,9 @@ import os
 import requests
 import time
 import urllib3
-import warnings
 
 from platformdirs import user_config_dir
 
-from pynsee.utils._get_token_from_insee import _get_token_from_insee
 from pynsee.utils.requests_params import _get_requests_session, _get_requests_headers, _get_requests_proxies
 
 
@@ -24,15 +22,14 @@ def opener(path, flags):
 
 
 def init_conn(
-    insee_key: str, insee_secret: str,
+    sirene_key: str, 
     http_proxy: str = "",
     https_proxy: str = ""
 ) -> None:
     """Save your credentials to connect to INSEE APIs, subscribe to api.insee.fr
 
     Args:
-        insee_key (str): user's key
-        insee_secret (str): user's secret
+        sirene_key (str): user's key for sirene API
         http_proxy (str, optional): Proxy server address, e.g. 'http://my_proxy_server:port'. Defaults to "".
         https_proxy (str, optional): Proxy server address, e.g. 'http://my_proxy_server:port'. Defaults to "".
 
@@ -41,44 +38,29 @@ def init_conn(
 
     Examples:
         >>> from pynsee.utils.init_conn import init_conn
-        >>> init_conn(insee_key="my_insee_key", insee_secret="my_insee_secret")
+        >>> init_conn(sirene_key="my_sirene_key")
         >>> #
         >>> # if the user has to use a proxy server use http_proxy and https_proxy arguments as follows:
         >>> from pynsee.utils.init_conn import init_conn
-        >>> init_conn(insee_key="my_insee_key",
-        >>>           insee_secret="my_insee_secret",
+        >>> init_conn(sirene_key="my_sirene_key",
         >>>           http_proxy="http://my_proxy_server:port",
         >>>           https_proxy="http://my_proxy_server:port")
         >>> #
         >>> # Alternativety you can use directly environment variables as follows:
         >>> # Beware not to commit your credentials!
         >>> import os
-        >>> os.environ['insee_key'] = 'my_insee_key'
-        >>> os.environ['insee_secret'] = 'my_insee_secret'
+        >>> os.environ['sirene_key'] = 'my_sirene_key'
         >>> os.environ['http_proxy'] = "http://my_proxy_server:port"
         >>> os.environ['https_proxy'] = "http://my_proxy_server:port"
     """
     logger.debug("SHOULD GET LOGGING")
-
-    try:
-        token = _get_token_from_insee(insee_key, insee_secret)
-    except Exception:
-        token = None
-
-    if token is None:
-        raise ValueError(
-            "!!! Could not generate token, please check that `insee_key` and "
-            "`insee_secret` are correct !!!"
-        )
-    else:
-        logger.info("Token has been created")
 
     proxies = _get_requests_proxies()
 
     queries = {
         "BDM": "https://api.insee.fr/series/BDM/V1/dataflow/FR1/all",
         "Metadata": "https://api.insee.fr/metadonnees/V1/codes/cj/n3/5599",
-        "Sirene": "https://api.insee.fr/entreprises/sirene/siret?q=activitePrincipaleUniteLegale:86.10*&nombre=1000",
+        "Sirene": "https://api.insee.fr/api-sirene/3.11/siret?q=activitePrincipaleUniteLegale:86.10*&nombre=1000",
         "Local Data": "https://api.insee.fr/donnees-locales/V0.1/donnees/geo-SEXE-DIPL_19@GEO2020RP2017/FE-1.all.all",
     }
 
@@ -98,16 +80,13 @@ def init_conn(
     for api, api_url in queries.items():
         headers = {
             "Accept": file_format[api],
-            "Authorization": "Bearer " + token,
             'User-Agent': user_agent['User-Agent']
         }
+        if api == "Sirene":
+            headers["X-INSEE-Api-Key-Integration"] = sirene_key        
 
-        with warnings.catch_warnings():
-            urllib3.disable_warnings(
-                urllib3.exceptions.InsecureRequestWarning)
-
-            results = session.get(
-                api_url, proxies=proxies, headers=headers, verify=False)
+        results = session.get(
+            api_url, proxies=proxies, headers=headers, verify=False)
         
         code = results.status_code
         
@@ -152,8 +131,7 @@ def init_conn(
 
     # save config
     config = {
-        "insee_key": insee_key,
-        "insee_secret": insee_secret,
+        "sirene_key": sirene_key,
         "http_proxy": http_proxy,
         "https_proxy": https_proxy
     }